move fin_* files to miopen directory
