+++ tests/test_fin_class.py
@@ -34,7 +34,7 @@ from dummy_machine import DummyMachine
 from tuna.tables import ConfigType
 from multiprocessing import Value, Lock, Queue
 from tuna.metadata import LOG_TIMEOUT
-from tuna.fin_class import FinClass
+from tuna.miopen.fin_class import FinClass
 
 sys.path.append("../tuna")
 sys.path.append("tuna")
