+++ tuna/go_fish.py
@@ -39,10 +39,10 @@ from tuna.metadata import MIOPEN_ALG_LIST
 from tuna.helper import print_solvers
 from tuna.miopen_tables import FinStep
 
-from tuna.fin_class import FinClass
+from tuna.miopen.fin_class import FinClass
 from tuna.utils.utility import get_env_vars
-from tuna.fin_builder import FinBuilder
-from tuna.fin_eval import FinEvaluator
+from tuna.miopen.fin_builder import FinBuilder
+from tuna.miopen.fin_eval import FinEvaluator
 from tuna.worker_interface import WorkerInterface
 from tuna.session import Session
