+++ tuna/miopen/fin_class.py
@@ -40,8 +40,8 @@ from tuna.worker_interface import WorkerInterface
 from tuna.dbBase.sql_alchemy import DbSession
 from tuna.metadata import FIN_CACHE
 from tuna.metadata import INVERS_DIR_MAP
-from tuna.fin_utils import compose_config_obj
-from tuna.fin_utils import get_fin_slv_status
+from tuna.miopen.fin_utils import compose_config_obj
+from tuna.miopen.fin_utils import get_fin_slv_status
 from tuna.config_type import ConfigType
 from tuna.utils.db_utility import session_retry
 from tuna.utils.db_utility import get_solver_ids, get_id_solvers
