+++ tuna/miopen/fin_eval.py
@@ -32,9 +32,9 @@ import json
 
 from sqlalchemy.exc import OperationalError
 
-from tuna.fin_class import FinClass
-from tuna.fin_utils import fin_job
-from tuna.fin_utils import get_fin_slv_status, get_fin_result
+from tuna.miopen.fin_class import FinClass
+from tuna.miopen.fin_utils import fin_job
+from tuna.miopen.fin_utils import get_fin_slv_status, get_fin_result
 from tuna.dbBase.sql_alchemy import DbSession
 from tuna.utils.db_utility import session_retry
